refactor: reset password stuff
